Update Tauto test requirement mappings for corrected EC suite
This is requirement sys-fw-0022, for which version 2 has the suite
corrected from faft_ec to faft_ec_fw_qual. This suite includes
everything in the former and more. This change uprevs existing mappings or adds new ones where needed.

Normally we would not replace an existing mapping, but this is a
subsystem integration in process with no users yet, so the old v01
mapping is just noise at this point.

BUG=b:284352086
TEST=none

Change-Id: I66f9f211a1aca41123f19ff7a97dbeda826df867
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4582094
Commit-Queue: Clayton Whitelaw <cawhitelaw@google.com>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Tested-by: Clayton Whitelaw <cawhitelaw@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>