tradefed_test: allow separate waiver yaml files to specify the same test case
Before this change, if the same test is specified in separate yaml
files, the file that gets read last will take priority.

We need this because certain tests are expected to fail on one set of
boards, while the same tests are to be marked as waived on another set
of boards.

BUG=b:286176301
TEST=cheets_CTS_R.internal.x86.all.CtsSettingsHostTestCases_-_CtsSyncManagerTestsCases

Change-Id: I123e8ba7a248af09f96ce5bb59e2639f377c165f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4596759
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>