bluetooth: fix original alias
It is possible that test_set_adapter_alias() sets a new alias

  1234567890123456789012345678901234567890...

and fails to reset it. This would make other tests, e.g.,
sa_valid_alias_test.floss, fail.

This patch resets the original alias to an empty string if
it is None to avoid the exceptional case.

BUG=b:286196433
TEST=The sa_valid_alias_test.floss test should pass in Testhaus.

Change-Id: Id8fee8e9c988311f682be952fa958d6f17f285e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4624187
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>