Disable full payload cellular AU tests
These tests fail consistently due to download the full payload taking
too long. To guarantee cellular is used for the update, we need to
disable the ethernet on the DUT. However autotest only allows the DUT to
be disconnected for 15 minutes before aborting the test. Changing this
to a configurable value is tricky and requires touching lots of autotest
internals.

These tests don't actually provide significant value though. We already
have plenty of test coverage for the difference between full and delta
payloads - the cellular update path we are testing is not concerned with
the payload type, only that the update medium is correctly determined
and cellular is used.

BUG=b:190648331
TEST=CQ

Change-Id: Ie099b6382d10173335dc3466d7a4ddd73be1f978
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4617118
Auto-Submit: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>