Add swarming IDs and BBIDs to provisioner.py cache requests
AU tests call into provisioner.py to do provisioning for m2n tests.
provisioner.py accepts an update_url from the AU tests, but it doesn't
use it directly so it's not that simple to just change what we're
passing in from the AU tests. Instead we can just change it in
provisioner.py directly at the site of the gs caching requests.

BUG=b:279489061
TEST=autoupdate_Basic.m2n.full

Change-Id: I894bc4275807dfe5db3713111a7a00c522b7dc3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4641869
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>