firmware_ECCharging: disable Charge Limit if it's enabled
This disables Charge Limit for this test, since it requires charging
unless the battery is full. Since it's a server side test, the
equivalent of the PowerPrefChanger was written via the host commandline.
Since firmware_ECCharging is the only server side autotest that appears
to need this, the code was added in the test.

BUG=b:284532839
TEST=firmware_ECCharging on atlas

Change-Id: I19fb6c6c32c73a107eb8109fee00414985069681
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4641879
Reviewed-by: Yanye Li <yanyeli@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>