autotest: catch exception in perf_monitor_service on missing keys
In perf_monitor_service, during post processing, it is possible
that some of the keys are present in the kernel perf data, due to
take of some of failure events during the test, this can cause
exception being generated when those data keys are queried e.g. Ip
data in snmp. This causes failure of the test, and can make the
perf test falky. As  a fix we should catch the keyError exception
and log the exception.

BUG=b:263978051
TEST=artificially added a new data key testKey in perf_monitor_graph
and ran the test, the change was able to catch the exception and log it.
05:01:09 INFO | autoserv| KeyException observed: missing key: 'testKey'

Change-Id: I150a68fff79473357652ce34985488ab2201ee7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4688415
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>