autotest: adjust KB backlight for some boards
For most boards, level == raw_percent/linear_percent
 == display_percent/scaled/nonlinear_percent/UI_percent.
But a few boards(kohaku and nightfury), the mentioned
equation doesn't apply because only part of the full
range of keyboard backlight pwm is used. And because
of this, some boards KB backlight is always set to be
below the min visible level, we want to correct this
behavior. When this happens, we set KB at lease to be
min level to reflect some power consumption.

BUG=b:291995701
TEST=run power_LoadTest.fast on a kohaku board

Change-Id: I983a329c286120fe4a41cfae2983b1caf70458c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4708136
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Jing Li <jingmuli@google.com>
Tested-by: Jing Li <jingmuli@google.com>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>