autotest: wifi: Add spectrum_mgmt_required for AttenPerf
The spectrum_mgmt_required=True AP option must be added, so the AP would
add Country Info in Beacon frames. This is needed for devices that use
self-managed wiphy which make use of the Country Info to update their
regulatory domain.

In the meantime, all devices are using country 00 which causes the device to use lower TX power and that leads to lower MCS index which causes  throughput to drop. For more information refer to comment b/285040925#comment36

BUG=b:285040925
TEST=Run the tests network_WiFi_AttenuatedPerf.*

Change-Id: I7db454496846651dbae40c706b55b5fe4b5cf320
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4725987
Reviewed-by: Kaidong Wang <kaidong@google.com>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>