bluetooth: select btpeers per device capabilities
This patch allows a test to specify the required capabilities
of a particular device type.

The test framework will select and run a btpeer with the required
capabilities.

If no such btpeers are found, the test raises TestNAError.

BUG=b:294168274
TEST=Run the aac test on a DUT with multiple btpeers.
The test should select one btpeer that has pipewire installed.
$ test_that --args "btpeer_host=$RASPBERRYPI_IP" $DUT_IP \
      bluetooth_AdapterAUHealth.au_a2dp_aac_test.floss

Change-Id: Ibbd5b3ba06d806e6d44d50f6bc3439e61d357e9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4996841
Reviewed-by: John Lai <johnlai@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>