bluetooth: select btpeers per device capabilities
This patch allows a test to specify the required capabilities
of a particular device type.

The test framework will select and run a btpeer with the required
capabilities.

If no such btpeers are found, the test raises TestNAError.

BUG=b:294168274
TEST=Run the aac test on a DUT with multiple btpeers.
The test should select one btpeer that has pipewire installed.
$ test_that --args "btpeer_host=$RASPBERRYPI_IP" $DUT_IP \
      bluetooth_AdapterAUHealth.au_a2dp_aac_test.floss

Change-Id: Ib7bcadf2281cf118e24435242d7e628e2d8508fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4817589
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: John Lai <johnlai@google.com>