autotest: Enable ADB without WiFi
This enables ADB to be used on a labstation without going over WiFi. It
acomplishes this by forwarding the ADB port from the phone to the
labstation over USB so the DUT/test_runner can instead bind to the
labstation IP address rather than the phone.

BUG=b:293875014
TEST=test_that localhost:2200 --args="phone_station=localhost
    phone_station_ssh_port=2201 android_serial=27191FDH2001SU
    phone_station_ip=192.168.1.99" tast.uwb

Change-Id: Id19495afc7cda10ed3c44adf6126382f80842692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4742494
Tested-by: Jason Stanko <jstanko@google.com>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Jason Stanko <jstanko@google.com>