faft: Update ECLidShutdown to handle slow kernels
The kernel on some pre-production devices may take a long time to power
down, so poll the DuT (with a timeout) until the desired power state is
is entered instead of waiting a fixed amount of time.

This improves the test's stability on devices where the kernel hasn't
been finalized.

BUG=b:294765831,b:194908032
TEST=FAFT firmware_ECLidShutdown passed on rex FW/OS 15579.0.0

Change-Id: Ib0a7ace7db76262a42414e4473f3a1167e42ac22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4781693
Tested-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Kshitij Shah <tij@google.com>
Commit-Queue: caveh jalali <caveh@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>