firmware_ConsecutiveBoot: Remove 1000+ iter runs from faft_stress suite
Currently, we have all versions of the ConsecutiveBoot tests running
in the faft_stress suite, but that's not necessary.  Removing the
really long running ones and keeping the 500 iteration version in.  As
stability improves we may try increasing the number of iterations in
the future.

BUG=b:297256735
TEST=None

Change-Id: I25ed46b6346951bb644ba5bee263e96d616597e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4809116
Commit-Queue: Shelley Chen <shchen@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>