firmware_WriteProtectFunc: Fix EC RO writeprotect assumption
The current flow of firmware_WriteProtectFunc looks like:
* Enable WP
* Read RO to _ro_before
* Generate a corrupted RO from _ro_before and name it _ro_test
* Use _ro_before and _ro_test to test WP enable
* Disable WP
* Use the same _ro_test to test WP disable

Based on the above flow, we could see that this test assumed that EC RO
with WP disable should be identical to EC RO with WP enable.
The assumption is totally wrong since there is a PSTATE flag inside RO
section. (b/172219703)

However, it was an unexpected coincidence that we always unconditionally
unlock the WP before reading the firmware. Therefore, the PSTATE we read
will always be `disabled` and thus this FAFT had passed for years.

But unconditionally unlock is really bad for flashrom since it
significantly slows down the flashrom itself due to WP operations and we
fixed that behavior in b/283779258 and CL:4676757.

So we need to read and corrupt the EC RO again after disabling WP. This
is the correct fix since we can't always make the assumption that EC RO
won't change despite the WP status.

BUG=b:283374161
TEST=Test on Geralt

Change-Id: Idb52680f9492c74345c53747f9e747e70189fe6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4809677
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Yidi Lin <yidilin@chromium.org>
Reviewed-by: Yidi Lin <yidilin@chromium.org>