buetooth: add test_audio_codec
The original A2DP test runs with the default SBC codec. This
patch checks the btsnoop log that AVDTP does set the correct
media codec in the configuration.

This will be useful to verify that the expected codec is used
when more tests are added to run A2DP tests with different codecs.

BUG=b:297812749
TEST=Run the a2dp tests.

$ test_that --args "btpeer_host=$RASPBERRYPI_IP" $DUT_IP \
  bluetooth_AdapterAUHealth.au_a2dp_test.floss
which should print something like
  INFO | autoserv| [>>> running: test_audio_codec]
  INFO | autoserv| codec configured: sbc (expected sbc)
  INFO | autoserv| [*** passed: test_audio_codec] elapsed_time: 3.652s

Change-Id: I681480d4b4299038905da703d4de5bf8f89d639e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4996839
Reviewed-by: John Lai <johnlai@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>