firmware_test: Flash EC RO only when restore_firmware
Due to b/282591058, when restore EC FW, `flashrom -p ec -w ec.bin`
doesn't work for ITE it83xx EC when the device is powered by
USB-C (e.g. Chromebox). The reason is EC needs to sysjump to RO
when it updates the RW FW, when the EC sysjumps to RO, the PD
contract needs to be renegotiated, which involves a PD hard reset.
So the EC briefly loses power and reboots.

The patch add a check and update EC RO only when battery is not
presented and EC chip is it83xx. In this case, it relies on EC
SW sync to restore EC RW FW.

BUG=b:301257525
TEST=Pass firmware_FWupdateThenSleep.wp on Dibbi which is powered
     by USB-C adapter

Change-Id: I7563062bcb1b2dd285d431bacf7c257fc6f76dee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4821152
Tested-by: Derek Huang <derekhuang@google.com>
Commit-Queue: Derek Huang <derekhuang@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>