wifi: Prevent router update on unsupported models
Adds a verification step to network_WiFi_UpdateRouter to check if the
router host is a ChromeOS router before making any changes. This test
being run on non-ChromeOS hosts causes them to break, as it wipes the
`/tmp` dir on them and also does not update them since this process
is not designed to support them. With this change the test will fail
before it clears the `/tmp` dir, rather than after.

BUG=b:296861897
TEST=cannot test locally as test does not support tunneled hostnames,
will rely on scheduled tests to verify (not a huge deal if this fails)

Change-Id: Ia6eacebd2efe9290d7d165d77f37fd99622380f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4852248
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>