wifi: Automatically calculate and report knee-point
This automatically calculates and reports the RvR knee points to
crosbolt. It includes a secondary run of throughput measurements at a
finer resolution around the knee point to ensure that we're accurately
resolving it. The individual refinement data points are not reported to
crosbolt but the improved final knee-point estimation is.

BUG=b:291746718
TEST=test_that --args="router_addr=localhost:2201 \
  pcap_addr=localhost:2202 atten_addr=localhost:2203" \
  localhost:2200 network_WiFi_AttenuatedPerf.ht40_ch001

Change-Id: Ia97448459bf9faa0aa9c5df797bb4ac638114e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4867884
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Jason Stanko <jstanko@google.com>
Commit-Queue: Jason Stanko <jstanko@google.com>