wifi: Add peripheral_wifi_state dep to unstable
The peripheral_wifi_state scheduling dimension is "WORKING" if all the
wifi routers in the testbed are in a good state, and "BROKEN"
otherwise. This adds a dependency to require that this is "WORKING"
to the control file which triggers the Tast wifi tests in the
wifi_matfunc_flaky and wifi_matfunc_openwrt_flaky suites to prevent
tests from being run on testbeds with known issues.

Once this is confirmed as working as expected in the lab, this
dependency will be added to the rest of the wifi tests' control files.

BUG=b:282840360
TEST=None, am unable to test locally. Will review in lab results
once merged.

Change-Id: Iee4c99ed4dcec4e4e418ac6e4fadfbd2bc438c4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4878325
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>