wifi: Add peripherial_wifi_state dep to all
The peripheral_wifi_state scheduling dimension is "WORKING" if all the
wifi routers in the testbed are in a good state, and "BROKEN"
otherwise. This adds a dependency to require that this is "WORKING"
to the remaining control files for all wifi tests that use routers.

BUG=b:282840360
TEST=None, am unable to test locally. Will review in lab results
once merged.

Change-Id: I6ad0a4ea0980e23a2fb0ee77cff3b806dd183411
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4878326
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Commit-Queue: Shijin Abraham <shijinabraham@google.com>
Tested-by: Shijin Abraham <shijinabraham@google.com>