power: Avoid pressing PWR_BTN when no display
When a Chromebox is not connected to an external display,
pressing power button will shut it down.

There is ongoing discussion on whether this behavior needs
to be changed, but even we were to change the signal handle
in Chrome, I would not expect this to land in short time.
So for now we just want to mark it an invalid wake source for
chromeboxes without external display.

BUG=b:259626865
TEST=test_that <fizz> power_WakeSources

Change-Id: I0232654aff785cd52b2a6053bac43119c10f162e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4888714
Tested-by: Yanye Li <yanyeli@google.com>
Commit-Queue: Yanye Li <yanyeli@google.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>