power_ProbeDriver: fix Charge Limit state change detection
power_ProbeDriver needs to disable Charge Limit before running, and the
code to detect when Charge Limit asynchronously disables was incorrect.
This fixes that code to correctly check that the battery is charging or
in a state when discharging is allowed.

BUG=b:217390814
TEST=power_ProbeDriver.probe_ac

Change-Id: I6d4082ee5d0b76ccbcfa3d9f8235b232b10a1a0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4890147
Reviewed-by: Yanye Li <yanyeli@google.com>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Yanye Li <yanyeli@google.com>