FWupdateThenSleep: apply --quirks=ec_partial_recovery=1
Due to b/282591058, we cannot update EC RW FW when the device is
with ITE it83xx EC and powered by USB-C (e.g. Dibbi Chromebox).
The reason is ITE EC needs to sysjump to RO when it updates the
RW FW, but when the EC sysjumps to RO, the PD contract needs to
be renegotiated, which involves a hard reset. So the EC briefly
loses power and reboots. For ITE EC Chromebox devices, we always
rely on EC software sync to update the EC RW FW.

This patch sets `--quirks=ec_partial_recovery=1` to update EC RO
only when the EC chip is it83xx and battery does not present.

BUG=b:302049459
TEST=Dibbi powered by USB-C adapter pass firmware_FWupdateThenSleep

Change-Id: Ied6bf48e635d639f23ae22388540ac23b6dbedad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4892548
Auto-Submit: Derek Huang <derekhuang@google.com>
Tested-by: Derek Huang <derekhuang@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Derek Huang <derekhuang@google.com>