Cr50CCDFirmwareUpdate: disable ite testing
i2c pseudo is not supported on the labstation. Disable the flash_ec part
of the test on duts that use iteflash until the labstation supports i2c
pseudo.

This patch can be reverted after i2c pseudo support is added to the
labstation.

BUG=b:302193501
TEST=run on ite device

Change-Id: Ic56709b7d904e8220d4aca86b14768354cab56fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4893905
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>