wifi: Increase wifi_matfunc timeout from 3h to 5h
We now have OpenWrt routers in our larger pools running the
wifi_matfunc suite, which take longer than Gale routers as they
reboot between tests. Due to the large size of the suite, this can
cause it to timeout with the current 3h limit. Our testing
with the wifi_matfunc_openwrt suite (which is a superset of
wifi_matfunc) found 5h to be a long enough timeout for OpenWrt routers.
Thus, this increases the regular wifi_matfunc suite timeout (and tast
run) to 5h to account for the fact that it is sometimes run with
OpenWrt routers.

Also corrected the METADATA for both control files to match other
wifi control files.

BUG=b:297059563
TEST=cannot be tested locally

Change-Id: I36cedb861ae159612a05ba10f25d3b2796f632b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4957383
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Shijin Abraham <shijinabraham@google.com>