power: Don't retry the wait for chgstate in EC
The intent of this sleep was to wait a few seconds
after DUT resuming such that EC can handle chgstate
command. This sleep should not stay inside the function
decorated by retry() because we only need it once.

Also increase it from 1 sec to 3 secs as it's occasionally
not enough for Octopus as shown in lab.

BUG=b:303528698
TEST=test_that <Octopus> power_WakeSources

Change-Id: I9019c450af66741281f65cdaba5b73beb8c243e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4984703
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Yanye Li <yanyeli@google.com>
Tested-by: Yanye Li <yanyeli@google.com>