power: Add option to not generate parallax report
This adds an option to not generate a parallax report for a power_Test.
Generating the parallax report sometimes takes as long as 20 minutes,
which may cause problems if we're trying to charge to a specific battery
percentage using the power_BatteryCharge test. This option to not
generate the report is enabled for power_BatteryChargeControl, which is
exlusively used for power_LW tests (only run in the lab, not PVS).

BUG=b:304627912
TEST=power_LW.power_LoadTest_caching

Change-Id: Ib6d75fcb983ccab1f15d9f34b2d1bb4ab7468819
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5085671
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yanye Li <yanyeli@google.com>