kernel_VerityCorruptRootfsA: Fully restore KERN-A
The test kernel_VerityCorruptRootfsA does the following:

1. Boot KERN-A
2. Corrupt KERN-A, verify DUT boots KERN-B
3. Restore KERN-A, verify DUT boots KERN-A

However, the restore step in (3) is not actually restoring KERN-A, so
the DUTs are continuing to boot KERN-B and failing the test.

The test puts the DUT into a good state during initialize() by calling
setup_kernel('a'). Rather than attempting to re-implement that logic,
just call it directly before verifying the device boots KERN-A in (3),
since we aren't attempting to validate that verity can be restored
correctly by the test, just that KERN-A is booted when verity is good.

This also means we can remove all of the backup/dump steps, since we
aren't attempting to re-use those values anymore, along with their
helper functions.

BUG=b:315211879
TEST=test_that --autotest_dir ~/chromiumos/src/third_party/autotest/files \
--args="servo_port=<port>" <IP> kernel_VerityCorruptRootfsA

Change-Id: Ibf77c6cacbacfc2deea7f8a6c747616efcb464c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5095950
Commit-Queue: Tim Van Patten <timvp@google.com>
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>