autoupdate_ForcedOOBEUpdate: Wait for progress after suspend/reboot
The test autoupdate_ForcedOOBEUpdate.interrupt.reboot.full does the
following:

1. Start an update
2. Interrupt the update with a device reboot
3. Verify the update resumes after the reboot

This works for the most part, but the test is a bit flaky and fails
with:

    The update was IDLE after interrupt.

This appears to be due to the test checking if the update was restarted
immediately once the DUT can be SSH'ed to. However, just because
networking and SSH are available, doesn't guarantee update_engine has
also had an opportunity to resume the update.

To handle this, poll for a bit after the reboot to give update_engine
time to resume the update. The test will now check for download progress
to increase before checking for idle.

This affects the suspend version of this test as well, so the polling is
performed for both cases:

autoupdate_ForcedOOBEUpdate.interrupt.suspend.full

The initial timeout is fairly high, since I'm not sure what sort of
expectation there is in terms of how long it takes an update to be
resumed after reboot. This can be reduced in the future when we have
something official to enforce.

BUG=b:315333897
TEST=test_that --autotest_dir ~/chromiumos/src/third_party/autotest/files \
--args="servo_port=<port>" <IP> autoupdate_ForcedOOBEUpdate.interrupt.reboot.full
TEST=test_that --autotest_dir ~/chromiumos/src/third_party/autotest/files \
--args="servo_port=<port>" <IP> autoupdate_ForcedOOBEUpdate.interrupt.suspend.full

Change-Id: I486781d3cb7ce74fa19a3de9a633ae177b576bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5099917
Tested-by: Tim Van Patten <timvp@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>