power_ProbeDriver: Fix Charge Limit disable condition
We were polling for the wrong condition for disabling Charge Limit. We
can't use battery_charging() as one of this conditions since that
returns true if the AC charger is online. This is the case for Charge
Limit. To avoid breaking tests that may rely on that behavior, this
instead used "not battery_discharging()", which only checks the battery
charging state.

BUG=b:315389658
TEST=power_ProbeDriver.probe_ac on volteer with Charge Limit enabled

Change-Id: I412a60c0f0ca90aeb5e10ea68bc6bf5f66de2ce7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5101284
Reviewed-by: Yanye Li <yanyeli@google.com>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>