Bluetooth: Start peers before resetting the bluetooth stack
One of the observed issue is that for chromebox in OOBE, if a peer was
emulated as LE HID before the test, when we start the peer it advertise
itself for a few seconds then if the test wants it to be a CL HID, in
the test the stack would find a device acts as LE only then after some
point it becomes BREDR only, which is not a common case we want the
stack to be able to deal with.

This CL simply change the order so that peers always emulate to the
desired device before we restart our stack.

BUG=b:312096641
TEST=run bluetooth_AdapterMDHealth

Change-Id: I7816ee2f5c589a3b26379cd297ffcbc2fa3b0a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5118320
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Tested-by: Yun-Hao Chung <howardchung@chromium.org>