SSH: Don't restart connection if RC = 0
Before this change, MainSsh.maybe_start() will restart the SSH
connection if "self._main_job.sp.poll() is not None". However, the SSH
background job immediately returns "0" when successful, leaving the SSH
connection open and usable. This means the SSH connection is constantly
regenerated once per second, flooding the logs and introducing
instability in the testing as the connection the test is using is
constantly changing.

The fix is to also check "self._main_job.sp.poll() != 0", and only
recreate the connection if there was an error.

BUG=b:317125390
TEST=test_that --autotest_dir ~/chromiumos/src/third_party/autotest/files \
     <IP> power_Display.fast

Change-Id: Ifb955703a92edc0df3d92cbe66a7babb0eb18fb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5138077
Reviewed-by: Mark Hasemeyer <markhas@google.com>
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>