wifi: Do not select ax tests in non-ax tast perf suite runs
The regular tast runs of the wifi_perf and wifi_perf_flaky suite are
not meant to select AX tests. This adds a negation of the tast
attribute that identifies the AX tests for those tast test selection
expressions, wificell_openwrt. The tests with that attribute will be
selected as part of wifi_perf_openwrt.

BUG=b:308616243
TEST=None

Change-Id: I55d8d0f44564f8256975a7787de8bffb9571bcd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5155432
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>