autotests: Remove StorageWearoutDetect from the suites
AFAIU the test was written for DUT's Storage audit purpose. Lab infra long
has separate tooling for the audit, and the test seems to be not useful for
the original purpose any more. Now it only generate unactionable noise for
the storage team. The test is not removed entirely yet for it is a part of
removable qual v1.It is not for any good reason, but given v1 should die
shortly when v3 is rolled out (removable skipped v2), mass removal of old
stuff will ensue afterwards.

BUG=b:317894681
TEST=none

Change-Id: I7a611822267f5bec6586af59a53dbd0354e821ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5162128
Reviewed-by: Otabek Kasimov <otabek@google.com>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>