wifi: Only run VerifyRouter on working gales
The network_WiFi_VerifyRouter test does not support OpenWrt
routers, but it is getting run against them and failing. This
adds a scheduling dependency to this test to only run it against
the testbeds with gale routers. The intent is to support all models
once this is migrated to Tast. Also adds the dependency requiring the
routers to be WORKING so that it does not get run on BROKEN testbeds.

BUG=b:297779631
TEST=None

Change-Id: I0169c2751b70996ff668f2905c31b72299e40b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5178136
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Commit-Queue: Jason Stanko <jstanko@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>