faft-gsc: remove testlab test from the suite
Remove the testlab test since we verify it on the dev board.

This test disables testlab mode. If the test fails to reenable it for
some reason, it can make lab devices harder to repair. It's kind of
risky to run this on lab devices. The testlab mode logic doesn't depend
on device hardware, so running on lab devices doesn't verify anything
the dev board already cover.

We can reenable the test if we find a reason to test on device hardware.

BUG=b:315976835
TEST=run devboard testlab mode test

Change-Id: I4410f93edfe27267c6426467d46380f81fd4bbd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5180715
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Auto-Submit: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Jett Rink <jettrink@google.com>
Tested-by: Mary Ruthven <mruthven@chromium.org>