Cr50FIPSDS: minor test fixes
Use poweroff to shutdown the device. It's what we use in the device
state tests. We know that works.

Check the EC console at the start of the test to verify it works ok.

Disable on ccd setups. The test verifies behavior through deep sleep.
This is tricky without servo micro and c2d2. Disable the test on ccd
setups.

Add more logging.

BUG=b:319253418
TEST=run on volteer

Change-Id: I25fb0de7c6ca8516fccd2b10bfe138cb4114441a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5180412
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Aadit Maheshkumar Modi <modia@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>