Bluetooth: Get the correct hci index for BluetoothRawSocket
It is known that the hci index for floss is not necessarily hci0.
However, the sockets in autotest is still hardcoded to hci0. This
causes some tests to fail even though BT still works like usual.

Add the support to get the active hci index, and supply the correct
index when querying data from BT socket.

BUG=b:316894815
TEST=Pass sa_adapter_suspend_resume_test.floss on RTL devices

Change-Id: I1961093bd6172388816385fa83fd220478cf875c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5183377
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>