tradefed_test: separate cache for DEV jobs under CFT
This change separates the TF cache for DEV jobs from others (LATEST,
DEV_WAIVER, ...), to limit the impact when a broken preview bundle on
ToT causes the entire cache to be wiped.

This change also adds the ability to set different sizes for different
caches (it used to be hardcoded to 100 GiB). To avoid taking up
additional drone disk space by introducing a separate cache, this
change:

* shrinks the size of SSP cache from 100 GiB to 50 GiB (all automated
  test jobs should run via CFT today)

* limits the size of the new CFT_DEV cache to 50 GiB so the total cache
  size doesn't change.

BUG=b:319223147
TEST=cheets_CTS_T.internal.arm.CtsAcceleration

Change-Id: I1e988ba09ab4afecacc5ddef85b065623dc1d044
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5227333
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Commit-Queue: Shao-Chuan Lee <shaochuan@chromium.org>