bluetooth: reduce RPA timeout to make RPA rotation
For LL privacy SR test, reduce the RPA timeout to force RPA rotation
at resume.

Return test NA if the wakeup or reconnect is not due to address
resolution in controller, because the test is invalid. The connection is
caused by advertising with public address due to race issue in peer.

BUG=b:322029301
TEST=sr_peer_wake_le_hid_with_rpa.floss

Change-Id: I567e1cfb7b5eac194d70b7ce15a66a39935e7cc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5238863
Tested-by: Zhengping Jiang <jiangzp@google.com>
Commit-Queue: Zhengping Jiang <jiangzp@google.com>
Reviewed-by: Yun-hao Chung <howardchung@chromium.org>