Increase session start timeout in autoupdate_Basic/login_LoginSuccess
autoupdate_Basic is flaking on some boards due to a timing issue in
login_LoginSuccess, where it takes longer than 10 seconds to receive the
'Session started' signal after logging in. Monitoring local test runs
shows that the login is actually complete and the UI is useable well
before the signal is received, so I think it's fine to increase this
timeout to improve stability.

BUG=b:321785429
TEST=autoupdate_Basic.m2n.full on coral

Change-Id: Iac241165d43daa8f0692a43ad984e3a042e6be17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5241883
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>