autotest: scan proc for lingering dmcrypt mounts
There is a suspicion that minijail held mounts can be a cause of issues with
cryptohome removal in some tests. It is not an issue in prod due to process
killer which would scan all process'es and kill those which are still holding
the mount. The triggered is tied to sign-off, which is not excercised in many
tests. Thus adding this logging to see what might be sporadically holding up
the mount in tests.

BUG=b:317866672
TEST=test_that 127.0.0.1:2222 platform_CryptohomeFio
TEST=check the debug result afterwards

Change-Id: Ia325506c9b7f4da008a62cc80d39e0e12ae78910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5240449
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>