autotest: Fix powerwash sticky kernel race condition
Post powerwashing, we MUST wait for the current boot to be marked
as the higher priority again. Otherwise, we will be racing with the
asynchronous task of marking the currently updated partition as
successful post invocation of a rollback.
However, simply check for the higher priority as well as success per
kernel boot priority will not work because the currently booted kernel
is already the highest priority.

BUG=b:324431261
TEST=CQ

Change-Id: I89e2888d0e382de105d989c191603a4336b4aa8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5281072
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>